bpf: Parse skb->data only once
Previously we parsed skb->data for twice: wan_egress/lan_ingress and
dae0peer_ingress. This is because the limit of bpf_sk_assign: we have to
call it within the netns where the socket is.

This patch manages to parse skb->data only once at
wan_egress/lan_ingress, where we leave a value in skb->cb[1] to tell
dae0peer_ingress:
1. if skb->cb[1] == TCP, then it's a new TCP conn, assign skb to TCP
   listener;
2. if skb->cb[1] == UDP, then it's a UDP, assign skb to UDP listener;
3. else it's an establised TCP conn, stack can take care of socket
   lookup;